requirements.txt
es_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.8.0/es_core_news_lg-3.8.0-py3-none-any.whl#sha256=7c6c212715a12f31aacde3361754436945ff7376fb24cde57d0c277c9c9b050b
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
joblib==1.5.3
nltk==3.9.3
numpy==2.4.3
python-dotenv==1.2.2
requests==2.32.5
scikit-learn==1.8.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
textual==8.1.1
